Add namespaces in configuration file
Maybe at some point the configuration could be placed in a file
shared with other tools.

+++ src/toolmaker/core.py
@@ -56,17 +56,18 @@ def _get_requirements(config):
     return requirements
 
 
-def _get_output_file_path(work_dir_path, tool_name, config):
-    output_dir_path = work_dir_path.joinpath(tool_name)
+def _get_output_file_path(work_dir_path, config):
+    output_dir_path = work_dir_path.joinpath(config['name'])
     output_file_name = config['output_file']
     output_file_path = output_dir_path.joinpath(output_file_name)
     return output_file_path
 
 
-def _build_pex(work_dir_path, tool_name, config, force):
+def _build_pex(work_dir_path, config, force):
     """ Build pex
     """
-    output_file_path = _get_output_file_path(work_dir_path, tool_name, config)
+    tool_name = config['name']
+    output_file_path = _get_output_file_path(work_dir_path, config)
     if force or not output_file_path.exists():
         LOGGER.info("Building pex tool '%s'...", tool_name)
         requirements = _get_requirements(config)
@@ -77,10 +78,11 @@ def _build_pex(work_dir_path, tool_name, config, force):
         LOGGER.info("Tool '%s' already exists, build skipped", tool_name)
 
 
-def _build_shiv(work_dir_path, tool_name, config, force):
+def _build_shiv(work_dir_path, config, force):
     """ Build shiv
     """
-    output_file_path = _get_output_file_path(work_dir_path, tool_name, config)
+    tool_name = config['name']
+    output_file_path = _get_output_file_path(work_dir_path, config)
     if force or not output_file_path.exists():
         LOGGER.info("Building shiv tool '%s'...", tool_name)
         requirements = _get_requirements(config)
@@ -91,10 +93,11 @@ def _build_shiv(work_dir_path, tool_name, config, force):
         LOGGER.info("Tool '%s' already exists, build skipped", tool_name)
 
 
-def _build_zapp(work_dir_path, tool_name, config, force):
+def _build_zapp(work_dir_path, config, force):
     """ Build zapp
     """
-    output_file_path = _get_output_file_path(work_dir_path, tool_name, config)
+    tool_name = config['name']
+    output_file_path = _get_output_file_path(work_dir_path, config)
     if force or not output_file_path.exists():
         LOGGER.info("Building zapp tool '%s'...", tool_name)
         requirements = _get_requirements(config)
@@ -105,34 +108,64 @@ def _build_zapp(work_dir_path, tool_name, config, force):
         LOGGER.info("Tool '%s' already exists, build skipped", tool_name)
 
 
+TOOL_SECTION_NAMES = (
+    'toolmaker.tool.pex',
+    'toolmaker.tool.shiv',
+    'toolmaker.tool.zapp',
+)
+
+
+def _parse_tool_config(raw_config, section_name):
+    tool_config = None
+    tokens = section_name.split(':', maxsplit=1)
+    if len(tokens) == 2 and tokens[0] in TOOL_SECTION_NAMES:
+        tool_config = {
+            'name': tokens[1],
+            'type': tokens[0].split('.')[2],
+        }
+        tool_config.update(raw_config[section_name])
+    return tool_config
+
+
+def parse_config(raw_config):
+    """ Parse the raw configration file to build a configuration dictionary.
+    """
+    config = {
+        'tools': {},
+    }
+    for section_name in raw_config.sections():
+        tool_config = _parse_tool_config(raw_config, section_name)
+        if tool_config:
+            config['tools'][tool_config['name']] = tool_config
+    return config
+
+
 def build(work_dir_path, config, tools_names, force=False):
     """ Build tools
     """
     LOGGER.info("Building tools %s...", tools_names)
 
     for tool_name in tools_names:
-        tool_config = config[tool_name]
-        if tool_name.endswith('.pex'):
-            _build_pex(
-                work_dir_path,
-                tool_name,
-                tool_config,
-                force,
-            )
-        if tool_name.endswith('.shiv'):
-            _build_shiv(
-                work_dir_path,
-                tool_name,
-                tool_config,
-                force,
-            )
-        if tool_name.endswith('.zapp'):
-            _build_zapp(
-                work_dir_path,
-                tool_name,
-                tool_config,
-                force,
-            )
+        if tool_name in config['tools']:
+            tool_config = config['tools'][tool_name]
+            if tool_config['type'] == 'pex':
+                _build_pex(
+                    work_dir_path,
+                    tool_config,
+                    force,
+                )
+            if tool_config['type'] == 'shiv':
+                _build_shiv(
+                    work_dir_path,
+                    tool_config,
+                    force,
+                )
+            if tool_config['type'] == 'zapp':
+                _build_zapp(
+                    work_dir_path,
+                    tool_config,
+                    force,
+                )
 
 
 def delete(work_dir_path, config, tools_names):
@@ -141,25 +174,29 @@ def delete(work_dir_path, config, tools_names):
     LOGGER.info("Deleting tools %s...", tools_names)
 
     for tool_name in tools_names:
-        output_file_path = _get_output_file_path(
-            work_dir_path,
-            tool_name,
-            config[tool_name],
-        )
-        if output_file_path.exists() and output_file_path.is_file():
-            LOGGER.info("Deleting file '%s'", output_file_path)
-            output_file_path.unlink()
-        output_dir_path = output_file_path.parent
-        if output_dir_path.exists() and output_dir_path.is_dir():
-            LOGGER.info("Deleting directory '%s'", output_dir_path)
-            try:
-                output_dir_path.rmdir()
-            except OSError as ose:
-                import errno
-                if ose.errno == errno.ENOTEMPTY:
-                    LOGGER.warning("Directory '%s' not empty", output_dir_path)
-                else:
-                    raise
+        if tool_name in config['tools']:
+            tool_config = config['tools'][tool_name]
+            output_file_path = _get_output_file_path(
+                work_dir_path,
+                tool_config,
+            )
+            if output_file_path.exists() and output_file_path.is_file():
+                LOGGER.info("Deleting file '%s'", output_file_path)
+                output_file_path.unlink()
+            output_dir_path = output_file_path.parent
+            if output_dir_path.exists() and output_dir_path.is_dir():
+                LOGGER.info("Deleting directory '%s'", output_dir_path)
+                try:
+                    output_dir_path.rmdir()
+                except OSError as ose:
+                    import errno
+                    if ose.errno == errno.ENOTEMPTY:
+                        LOGGER.warning(
+                            "Directory '%s' not empty",
+                            output_dir_path,
+                        )
+                    else:
+                        raise
 
 
 # EOF
